refactor: Move user store to context and update component props for better state management
